Add LocalManager and refactor InstallLocalPackageCommand
